+++ vedo/plotter.py
@@ -77,7 +77,6 @@ def show(*actors, **options):
     :param float azimuth/elevation/roll:  move camera accordingly
     :param str viewup:  either ['x', 'y', 'z'] or a vector to set vertical direction
     :param bool resetcam:  re-adjust camera position to fit objects
-
     :param dict camera: Camera parameters can further be specified with a dictionary
         assigned to the ``camera`` keyword (E.g. `show(camera={'pos':(1,2,3), 'thickness':1000,})`)
 
@@ -113,6 +112,7 @@ def show(*actors, **options):
 
     :param bool interactive:  pause and interact with window (True)
         or continue execution (False)
+
     :param float rate:  maximum rate of `show()` in Hertz
     :param int interactorStyle: set the type of interaction
 
@@ -126,15 +126,13 @@ def show(*actors, **options):
         - 7 = RubberBandZoom
         - 8 = Context
         - 9 = 3D
-        -10 = Terrain
-        -11 = Unicam
+        - 10 = Terrain
+        - 11 = Unicam
 
     :param bool q:  force program to quit after `show()` command returns.
 
     :param bool new: if set to `True`, a call to ``show`` will instantiate
         a new ``Plotter`` object (a new window) instead of reusing the first created.
-
-        See e.g.: |readVolumeAsIsoSurface.py|_
 
     :return: the current ``Plotter`` class instance.
 
@@ -1203,49 +1201,7 @@ class Plotter:
             - 12, show polar axes
             - 13, draw a simple ruler at the bottom of the window
 
-        Axis type-1 can be fully customized by passing a dictionary ``axes=dict()`` where:
-
-            - `xtitle`,                ['x'], x-axis title text
-            - `xrange`,               [None], x-axis range in format (xmin, ymin), default is automatic.
-            - `numberOfDivisions`,    [None], approximate number of divisions on the longest axis
-            - `axesLineWidth`,           [1], width of the axes lines
-            - `gridLineWidth`,           [1], width of the grid lines
-            - `reorientShortTitle`,   [True], titles shorter than 2 letter are placed horizontally
-            - `originMarkerSize`,     [0.01], draw a small cube on the axis where the origin is
-            - `titleDepth`,              [0], extrusion fractional depth of title text
-            - `xyGrid`,               [True], show a gridded wall on plane xy
-            - `yzGrid`,               [True], show a gridded wall on plane yz
-            - `zxGrid`,               [True], show a gridded wall on plane zx
-            - `zxGrid2`,             [False], show zx plane on opposite side of the bounding box
-            - `xyGridTransparent`    [False], make grid plane completely transparent
-            - `xyGrid2Transparent`   [False], make grid plane completely transparent on opposite side box
-            - `xyPlaneColor`,       ['gray'], color of the plane
-            - `xyGridColor`,        ['gray'], grid line color
-            - `xyAlpha`,              [0.15], grid plane opacity
-            - `xyFrameLine`,          [None], add a frame for the plane
-            - `showTicks`,            [True], show major ticks
-            - `xTitlePosition`,       [0.32], title fractional positions along axis
-            - `xTitleOffset`,         [0.05], title fractional offset distance from axis line
-            - `xTitleJustify`, ["top-right"], title justification
-            - `xTitleRotation`,          [0], add a rotation of the axis title
-            - `xLineColor`,      [automatic], color of the x-axis
-            - `xTitleColor`,     [automatic], color of the axis title
-            - `xTitleBackfaceColor`,  [None],  color of axis title on its backface
-            - `xTitleSize`,          [0.025], size of the axis title
-            - 'xTitleItalic',            [0], a bool or float to make the font italic
-            - `xHighlightZero`,       [True], draw a line highlighting zero position if in range
-            - `xHighlightZeroColor`, [autom], color of the line highlighting the zero position
-            - `xTickLength`,         [0.005], radius of the major ticks
-            - `xTickThickness`,     [0.0025], thickness of the major ticks along their axis
-            - `xTickColor`,      [automatic], color of major ticks
-            - `xMinorTicks`,             [1], number of minor ticks between two major ticks
-            - `xPositionsAndLabels`       [], assign custom tick positions and labels [(pos1, label1), ...]
-            - `xLabelPrecision`,         [2], nr. of significative digits to be shown
-            - `xLabelSize`,          [0.015], size of the numeric labels along axis
-            - `xLabelOffset`,        [0.025], offset of numeric labels
-            - 'xFlipText'.           [False], flip axis title and numeric labels orientation
-            - `tipSize`,              [0.01], size of the arrow tip
-            - `limitRatio`,           [0.04], below this ratio don't plot small axis
+        Axis type-1 can be fully customized by passing a dictionary ``axes=dict()``.
 
             :Example:
 
@@ -1260,7 +1216,9 @@ class Plotter:
                                  }
                     )
 
-        |customAxes| |customAxes.py|_
+        |customAxes1| |customAxes1.py|_  |customAxes2.py|_ |customAxes3.py|_
+
+        |customIndividualAxes| |customIndividualAxes.py|_
         """
         addons.addGlobalAxes(axtype, c)
         return self
@@ -1317,7 +1275,7 @@ class Plotter:
         :param str viewup:  either ['x', 'y', 'z'] to set vertical direction
         :param bool resetcam:  re-adjust camera position to fit objects
         :param dict camera: Camera parameters can further be specified with a dictionary assigned
-           to the ``camera`` keyword (E.g. `show(camera={'pos':(1,2,3), 'thickness':1000,})`)
+            to the ``camera`` keyword (E.g. `show(camera={'pos':(1,2,3), 'thickness':1000,})`)
 
             - pos, `(list)`,  the position of the camera in world coordinates
             - focalPoint `(list)`, the focal point of the camera in world coordinates
@@ -1360,10 +1318,11 @@ class Plotter:
             - 7 = RubberBandZoom
             - 8 = Context
             - 9 = 3D
-            -10 = Terrain
-            -11 = Unicam
+            - 10 = Terrain
+            - 11 = Unicam
 
         :param bool q:  force program to quit after `show()` command returns.
+
         """
         at = options.pop("at", None)
         axes = options.pop("axes", settings.defaultAxesType)
@@ -1825,6 +1784,7 @@ class Plotter:
         :param int at: specify the renderer number
         :param pos: icon position in the range [1-4] indicating one of the 4 corners,
                     or it can be a tuple (x,y) as a fraction of the renderer size.
+
         :param float size: size of the square inset.
         :param bool draggable: if True the subrenderer space can be dragged around.
         :param c: color of the inset frame when dragged