+++ vedo/pyplot.py
@@ -1745,7 +1745,7 @@ def _histogramHexBin(
             else:
                 p = (i / ki, j / kj + 0.45, 0)
             q = (p[0] / n * 1.2 * dx + xmin, p[1] / m * dy + ymin, 0)
-            ids = cloud.closestPoint(q, radius=r, returnIds=True)
+            ids = cloud.closestPoint(q, radius=r, returnCellId=True)
             ne = len(ids)
             if fill:
                 t.Translate(p[0], p[1], ne / 2)
@@ -1972,7 +1972,7 @@ def _histogramSpheric(
     cntrs = sg.cellCenters()
     counts = np.zeros(len(cntrs))
     for p in ptsvals:
-        cell = sg.closestPoint(p, returnIds=True)
+        cell = sg.closestPoint(p, returnCellId=True)
         counts[cell] += 1
     acounts = np.array(counts)
     counts *= (rmax - 1) / np.max(counts)